refactored prior to more analysis for Rib Tables and BGP attributes
